Upgrade babelfishpg_money support to PG-13
The source code for babelfishpg_money was lifted from a
version of 2nd-Quadrant's fixeddecimal code, that only
supported PostgreSQL up to version 12 at that time. This
patch upgrades the support to 13, which enables
PARALLEL SAFE for all functions, adds cross type operators
for int8 and the operator class for BRIN indexes.