Use locally staged but not committed source changes
Allow locally added but not yet committed changes in
the engine and extensions subprojects. The way this works
is to make changes and do 'git add' (but not commit).
The build script will then create a stash for the tarballs
instead of the tag/branch specified in the build.conf file.